1-3 points done

+++ app/src/main/java/ru/tinkoff/ru/seminar/model/ForecastEntity/Main.java
@@ -28,7 +28,7 @@ public class Main {
     private int humidity;
     @SerializedName("temp_kf")
     @Expose
-    private int tempKf;
+    private double tempKf;
 
     public double getTemp() {
         return temp;
@@ -86,7 +86,7 @@ public class Main {
         this.humidity = humidity;
     }
 
-    public int getTempKf() {
+    public double getTempKf() {
         return tempKf;
     }
